feat: Add hyprland wallpaper, idle, logout, and lock functionality
This commit adds the necessary code changes to introduce the hyprland wallpaper, idle, logout, and lock functionality. The `hyprpaper.nix` file is created to configure the hyprland wallpaper daemon, `hypridle.nix` is added to enable screen idle functionality, `wlogout.nix` is included to customize the logout screen, and `hyprlock.nix` is introduced to configure the lock screen. These changes enhance the user experience and provide a more cohesive and visually appealing interface.