feat: Update hyprlock.nix to use lib.mkForce for enable flag
This commit modifies the `hyprlock.nix` file to use `lib.mkForce` for the `enable` flag in the `programs.hyprlock` section. By using `lib.mkForce true`, we ensure that the `enable` flag is always set to true, regardless of the value in the configuration. This change improves the reliability and consistency of the `hyprlock` program.