chore: Update docker default.nix to allow network interface in firewall
This commit updates the `default.nix` file in the `docker` module to include the `networkInterface` option. This option allows specifying the network interface to allow in the firewall. By default, it is set to `docker0`. This change enhances the flexibility and security of the Docker setup by allowing specific network interfaces to be whitelisted in the firewall configuration.