chore: Enable Mullvad IP exclusion in work default.nix
This commit enables the Mullvad IP exclusion feature in the work default.nix file. By setting the `enableExludeIPs` flag to true and specifying the excluded IP addresses, we ensure that these IPs are excluded from the Mullvad VPN connection. This change enhances the security and privacy of the work environment.